Atualização de repository (#60)

+++ src/main/java/cetam/projeto01grupo05/repository/DevolucaoRepository.java
@@ -7,6 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface DevolucaoRepository extends JpaRepository<Devolucao, Integer> {
+public interface DevolucaoRepository extends JpaRepository<Devolucao, Long> {
+
     Optional<Devolucao> findByEmprestimoIdEmprestimo(Long idEmprestimo);
 }
+
